Improve the accuracy of displayed prices and totals throughout the platform
Modify price calculations to use integers and round the totals for more precise currency representation.

+++ client/src/components/ProductCard.tsx
@@ -44,7 +44,7 @@ export function ProductCard({ product, viewMode = "grid" }: ProductCardProps) {
   };
 
   const formatPrice = (price: string) => {
-    return `TZS ${parseFloat(price).toLocaleString()}`;
+    return `TZS ${parseInt(price).toLocaleString()}`;
   };
 
   if (viewMode === "list") {
